refactor : move RestResponse to models/io
