+++ src/apis/auth/mutations/useAuthLoginMutation.ts
@@ -1,7 +1,8 @@
 import { useMutation } from "@tanstack/react-query";
 
 import { AuthLoginRequest } from "@apis/auth/models/AuthLogin";
-import { RestErrorResponse, RestResponse } from "@apis/_models/RestResponse";
+
+import { RestErrorResponse, RestResponse } from "@models/io/RestResponse";
 
 import rest from "@utils/rest";
